fix: renaming folders
