+++ src/features/BuildingJourney/components/BuildingJourneyPart/BuildingJourneyPart.tsx
