+++ src/features/BuildingJourney/components/HistoryCard/HistoryCardImage.tsx
