+++ src/features/DesignStory/components/DesignStoryConcept/DesignStoryConceptCard1.tsx
