+++ src/features/DesignStory/components/DesignStoryConcept/DesignStoryConceptCard2.tsx
