+++ src/features/DesignStory/components/DesignStoryConcept/DesignStoryConceptCard3.tsx
