+++ src/features/DesignStory/components/DesignStoryConcept/DesignStoryConceptCard4.tsx
