+++ src/features/DesignStory/components/DesignStoryConcept/DesignStoryConceptCard6.tsx
