+++ src/features/DesignStory/components/DesignStoryConcept/DesignStoryConceptCard7.tsx
