+++ src/features/DesignStory/components/DesignStoryConcept/DesignStoryConceptCard9.tsx
