+++ src/features/DesignStory/components/DesignStoryConcept/DesignStoryConceptFinal.tsx
