+++ src/features/DesignStory/components/DesignStoryConcept/index.ts
