+++ src/features/DesignStory/components/DesignStoryGallery/DesignStoryGalleryPhoto.tsx
