+++ src/features/DesignStory/components/DesignStoryIntro/DesignStoryIntroCard1.tsx
