+++ src/features/DesignStory/components/DesignStoryIntro/DesignStoryIntroCard2.tsx
